cpufreq: interactive governor drops bits in time calculation, fix it

+++ drivers/cpufreq/cpufreq_interactive.c
@@ -454,8 +454,8 @@ static u64 update_load(int cpu)
 	struct cpufreq_interactive_cpuinfo *pcpu = &per_cpu(cpuinfo, cpu);
 	u64 now;
 	u64 now_idle;
-	unsigned int delta_idle;
-	unsigned int delta_time;
+	u64 delta_idle;
+	u64 delta_time;
 	u64 active_time;
 #if defined(CONFIG_SEC_PM) || defined(CONFIG_MODE_AUTO_CHANGE)
 	unsigned int cur_load = 0;
@@ -465,8 +465,8 @@ static u64 update_load(int cpu)
 #endif
 
 	now_idle = get_cpu_idle_time(cpu, &now);
-	delta_idle = (unsigned int)(now_idle - pcpu->time_in_idle);
-	delta_time = (unsigned int)(now - pcpu->time_in_idle_timestamp);
+	delta_idle = (now_idle - pcpu->time_in_idle);
+	delta_time = (now - pcpu->time_in_idle_timestamp);
 
 	if (delta_time <= delta_idle)
 		active_time = 0;
